finish dbcontext , repository,add migration,add category command and query and validation,

+++ Blog.Domain/Entities/Categories/Category.cs
@@ -24,7 +24,7 @@ namespace Blog.Domain.Entities.Categories
         public string Description { get; set; }
         public string Image { get; set; }
         public string Icon { get; set; }
-        public int? ParentId { get; set; }
+        public long? ParentId { get; set; }
         public bool ShowInHomePage { get; set; }
         public int DisplayOrder { get; set; }
         #endregion
